feat: attempt to add transfers to cap gain calc
